move Error to CommandQuery
