+++ src/CommandQuery.AspNetCore/_InternalsVisibleTo.cs
