+++ src/CommandQuery/Error.cs
